add multistack subpackage
