+++ bigfish/segmentation/__init__.py
@@ -23,7 +23,13 @@ from .postprocess import label_instances
 from .postprocess import merge_labels
 from .postprocess import clean_segmentation
 from .postprocess import remove_disjoint
-from .postprocess import match_nuc_cell
+from .postprocess import center_mask_coord
+from .postprocess import from_boundaries_to_surface
+from .postprocess import from_surface_to_boundaries
+from .postprocess import from_binary_to_coord
+from .postprocess import complete_coord_boundaries
+from .postprocess import from_coord_to_frame
+from .postprocess import from_coord_to_surface
 
 from .utils import thresholding
 from .utils import compute_mean_diameter
@@ -53,7 +59,13 @@ _postprocess = [
     "merge_labels",
     "clean_segmentation",
     "remove_disjoint",
-    "match_nuc_cell"]
+    "center_mask_coord",
+    "from_boundaries_to_surface",
+    "from_surface_to_boundaries",
+    "from_binary_to_coord",
+    "complete_coord_boundaries",
+    "from_coord_to_frame",
+    "from_coord_to_surface"]
 
 _utils = [
     "thresholding",
